lots of relatively boring Size/UInt/Int/SignedSize issues
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@4855 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/MAPMATCHING/PoseClusteringShiftSuperimposer.h
@@ -43,13 +43,13 @@ namespace OpenMS
 {
   /**
 		@brief Superimposer that uses a voting scheme to find a good translation.
-		
+
 		It works similarly to the @ref PoseClusteringAffineSuperimposer , but only shifts are
 		considered. Scaling is not corrected.
-		
+
 		@todo remove m/z correction and parameters; consider full rewrite (Clemens)
-		
-		@htmlinclude OpenMS_PoseClusteringShiftSuperimposer.parameters    
+
+		@htmlinclude OpenMS_PoseClusteringShiftSuperimposer.parameters
 
 		@ingroup MapAlignment
   */
@@ -63,7 +63,7 @@ namespace OpenMS
 	    	: BaseSuperimposer()
 	    {
 				setName(getProductName());
-				
+
 	      defaults_.setValue("input_map:bucket_size:RT",150.0,"Number of surrounding buckets of element indices to be considered when computing shifts.",StringList::create("advanced"));
 	      defaults_.setValue("input_map:bucket_size:MZ",4.0,"Number of surrounding buckets of element indices to be considered when computing shifts.",StringList::create("advanced"));
 	      defaults_.setValue("transformation_space:shift_bucket_size:RT",5.0,"Defines the shift parameter's bucket size during histograming.");
@@ -73,29 +73,29 @@ namespace OpenMS
 	      defaults_.setValue("transformation_space:bucket_window_shift:RT",2,"Number of surrounding buckets of shift indices to be considered when computing shifts.",StringList::create("advanced"));
 	      defaults_.setValue("transformation_space:bucket_window_shift:MZ",1,"Number of surrounding buckets of shift indices to be considered when computing shifts.",StringList::create("advanced"));
 				subsections_.push_back("debug");
-				
+
 	      defaultsToParam_();
 	    }
-	
+
 	    /// Destructor
 	    virtual ~PoseClusteringShiftSuperimposer()
 	    {
 	    }
-	
+
 	    /**
 	    	@brief Estimates the transformation and fills the given mapping function
-	    	
+
 	    	@note Exactly two input maps must be given.
-	    	
+
 	    	@exception IllegalArgument is thrown if the input maps are invalid.
 	    */
 	    virtual void run(const std::vector<ConsensusMap>& maps, std::vector<TransformationDescription>& transformations)
 	    {
 	    	if (maps.size()!=2) throw Exception::IllegalArgument(__FILE__,__LINE__,__PRETTY_FUNCTION__,"Excactly two input maps are required");
-	    	
+
 	    	model_map_ = &(maps[MODEL]);
 	    	scene_map_ = &(maps[SCENE]);
-	    	
+
 	      // clear the member
 	      element_bucket_[Peak2D::RT].clear();
 	      element_bucket_[Peak2D::MZ].clear();
@@ -109,19 +109,19 @@ namespace OpenMS
 	      computeShiftBuckets_();
 	      computeShift_(transformations[0]);
 	    }
-	
+
 	    /// Returns an instance of this class
 	    static BaseSuperimposer* create()
 	    {
 	      return new PoseClusteringShiftSuperimposer();
 	    }
-	
+
 	    /// Returns the name of this module
 	    static const String getProductName()
 	    {
 	      return "poseclustering_shift";
 	    }
-	    
+
 	  protected:
 
 	    /// Internal representation of a shift used in PoseClusteringShiftSuperimposer
@@ -133,7 +133,7 @@ namespace OpenMS
 				    quality(0.0)
 				{
 				}
-				
+
 	      DPosition<2> position;
 	      DoubleReal quality;
 	    };
@@ -145,14 +145,14 @@ namespace OpenMS
 	    typedef Matrix < DoubleReal > ShiftQualityMatrixType;
 	    typedef Matrix < Shift > ShiftMatrixType;
 			//@}
-	  	
+
 	  	///map index names
 	  	enum
 			{
 				MODEL=0,
 				SCENE=1
 			};
-	  	
+
 	    virtual void updateMembers_()
 	    {
 	      shift_bucket_size_[0] = (DoubleReal)param_.getValue("transformation_space:shift_bucket_size:RT");
@@ -170,19 +170,19 @@ namespace OpenMS
 	    {
 	      // Shorthands ...
 	      DPosition<2> & fbs = element_bucket_size_;
-	
+
 				const ConsensusMap* map_array[2] = {model_map_, scene_map_};
-	
+
 	      for ( UInt map_index = 0; map_index < 2; ++map_index )
 	      {
 	        // Shorthands ...
 	      	ConsensusMap const & fm = *(map_array[map_index]);
 	        DBoundingBox<2> & fmpbb  = element_map_position_bounding_box_[map_index] ;
 	        DBoundingBox<1> & fmibb  = element_map_intensity_bounding_box_[map_index];
-	
+
 	        fmpbb.clear();
 	        fmibb.clear();
-	
+
 	        // Compute the bounding box for the element map, with respect to
 	        // position and intensity.
 	        for ( ConsensusMap::const_iterator fm_iter = fm.begin();
@@ -194,7 +194,7 @@ namespace OpenMS
 	          fmibb.enlarge(fm_iter->getIntensity());
 	        }
 	      }
-	
+
 	      // Next we will enlarge each element_map_position_bounding_box_ such
 	      // that all buckets will have the same diagonal.  To provide against
 	      // rounding errors, we allocate one bucket more than needed (in each
@@ -206,7 +206,7 @@ namespace OpenMS
 	        DBoundingBox<2> const & fmpbb  = element_map_position_bounding_box_[map_index] ;
 	        DBoundingBox<2> & fmpbbe = element_map_position_bounding_box_enlarged_[map_index] ;
 	        ElementBucketMatrixType       & fb     = element_bucket_[map_index];
-	
+
 	        // Compute num_buckets.  Compute extra margin to make bounding box a
 	        // multiple of element buckets.
 	        DPosition<2> const diagonal = fmpbb.diagonal();
@@ -217,19 +217,19 @@ namespace OpenMS
 	          num_buckets[dimension] = int(1.1 + diagonal[dimension]/fbs[dimension]);
 	          diagonal_enlarged[dimension] = fbs[dimension] * num_buckets[dimension];
 	        }
-	
+
 	        // The extra margin.
 	        DPosition<2> extra_element_bucket_size_(diagonal_enlarged-diagonal);
 	        extra_element_bucket_size_ /= 2;
-	
+
 	        // Compute the enlarged element map bounding box accordingly.
 	        fmpbbe.clear();
 	        fmpbbe.enlarge( fmpbb.min() - extra_element_bucket_size_ );
 	        fmpbbe.enlarge( fmpbb.max() + extra_element_bucket_size_ );
-	
+
 	        // Resize element_bucket_[map_index] accordingly.
 	        fb.resize(num_buckets[Peak2D::RT],num_buckets[Peak2D::MZ]);
-	
+
 	        // Now, finally, we store the indices of the elements in their
 	        // corresponding buckets.
 	        DPosition<2> const & fmpbbe_min = fmpbbe.min();
@@ -238,7 +238,7 @@ namespace OpenMS
 	          DPosition<2> position = fm[index].getPosition() - fmpbbe_min;
 	          fb ( UInt(position[Peak2D::RT]/fbs[Peak2D::RT]), UInt(position[Peak2D::MZ]/fbs[Peak2D::MZ]) ).push_back(index);
 	        }
-	
+
 	        // Optionally, write debug output as specified in param.
 	        if ( getParameters().exists("debug:feature_buckets_file") )
 	        {
@@ -263,10 +263,10 @@ namespace OpenMS
 	        }
 	      }
 	    }
-	
+
 	    /**
 	    	@brief Fill the buckets of shifts.
-	
+
 		    Note that computeElementBuckets_() must have been called before to make this work properly.
 	    */
 	    void computeShiftBuckets_()
@@ -277,7 +277,7 @@ namespace OpenMS
 	      DBoundingBox<2> & tbb = shift_bounding_box_ ;
 	      DBoundingBox<2> & tbbe = shift_bounding_box_enlarged_ ;
 	      UInt const (&fbw)[2] = element_bucket_window_;
-	
+
 	      // Compute the bounding box for the shift map
 	      {
 	        tbb.clear();
@@ -290,17 +290,17 @@ namespace OpenMS
 	      // have the same diagonal.  To provide against rounding errors, we
 	      // allocate one bucket more than needed (in each dimension) and shift
 	      // the grid by one-half.
-	
+
 	      DPosition<2> half_of_shift_bucket_size_(tbs);
 	      half_of_shift_bucket_size_ /= 2;
-	
+
 	      // Adjust the enlarged shift map bounding box accordingly.
 	      {
 	        tbbe.clear();
 	        tbbe.enlarge( tbb.min() - half_of_shift_bucket_size_ );
 	        tbbe.enlarge( tbb.max() + half_of_shift_bucket_size_ );
 	      }
-	
+
 	      // Compute shift_bucket_size_ and num_buckets.
 	      DPosition<2> diagonal = tbbe.diagonal();
 	      int num_buckets[2];
@@ -309,17 +309,17 @@ namespace OpenMS
 	        num_buckets[dimension] = int(diagonal[dimension]/tbs[dimension]);
 	        tbs[dimension] = diagonal[dimension] / num_buckets[dimension];
 	      }
-	
+
 	      // Resize shift_bucket_ accordingly.
 	      tb.resize(num_buckets[Peak2D::RT]+1,num_buckets[Peak2D::MZ]+1);
-	
+
 	      // Clear the shift buckets.
 	      std::fill(tb.begin(),tb.end(),DoubleReal(0));
-	
-	
+
+
 	      // Resize shift_matrix_ according to element_bucket_[Peak2D::MZ]
 	      //         tm.resize(element_bucket_[Peak2D::MZ].sizePair());
-	
+
 	      // Now we store the shifts for all relevant element pairs in their
 	      // corresponding buckets.  Each shift is distributed among its
 	      // four neighboring "buckets", with weights according to the distances
@@ -330,23 +330,23 @@ namespace OpenMS
 	      // way we can associate the shifts vectors to buckets of the
 	      // image, and when we will later apply it, we will not change the
 	      // pre-image, which might be a consensus or so.
-	
+
 	      // progress dots
 	      Int progress_dots = 0;
 				if (this->param_.exists("debug::progress_dots"))
 				{
 	      	progress_dots = (Int)this->param_.getValue("debug:progress_dots");
 			 	}
-	
+
 	      DPosition<2> const & tbbe_min = tbbe.min();
-	
+
 	      // Compute the index shift of corresponding element buckets of model and scene.
 	      DPosition<2> const fmpbbe_min_offset =
 	        element_map_position_bounding_box_enlarged_[SCENE].min() -
 	        element_map_position_bounding_box_enlarged_[MODEL].min();
 	      int const element_buckets_index_offset_RT = int ( fmpbbe_min_offset[Peak2D::RT] / element_bucket_size_[Peak2D::RT] );
 	      int const element_buckets_index_offset_MZ = int ( fmpbbe_min_offset[Peak2D::MZ] / element_bucket_size_[Peak2D::MZ] );
-	
+
 	      // iterate over buckets of scene
 	      for ( UInt scene_bucket_index_RT = 0;
 	            scene_bucket_index_RT < element_bucket_[SCENE].rows();
@@ -358,23 +358,23 @@ namespace OpenMS
 	              ++scene_bucket_index_MZ
 	            )
 	        {
-	
+
 	          // compute the corresponding bucket in the model
 	          int const model_bucket_index_center_RT = scene_bucket_index_RT + element_buckets_index_offset_RT;
 	          int const model_bucket_index_center_MZ = scene_bucket_index_MZ + element_buckets_index_offset_MZ;
-	
+
 	          // iterate over buckets of model
 	          for ( int model_bucket_index_RT
 	                =  std::max<int>( model_bucket_index_center_RT - fbw[Peak2D::RT], 0 );
 	                model_bucket_index_RT
-	                <= std::min<int>( model_bucket_index_center_RT + fbw[Peak2D::RT], element_bucket_[MODEL].rows()-1 );
+	                <= std::min<int>( model_bucket_index_center_RT + fbw[Peak2D::RT], (int)element_bucket_[MODEL].rows()-1 );
 	                ++model_bucket_index_RT
 	              )
 	          {
 	            for ( int model_bucket_index_MZ
 	                  =  std::max<int>( model_bucket_index_center_MZ - fbw[Peak2D::MZ], 0 );
 	                  model_bucket_index_MZ
-	                  <= std::min<int>( model_bucket_index_center_MZ + fbw[Peak2D::MZ], element_bucket_[MODEL].cols()-1 );
+	                  <= std::min<int>( model_bucket_index_center_MZ + fbw[Peak2D::MZ], (int)element_bucket_[MODEL].cols()-1 );
 	                  ++model_bucket_index_MZ
 	                )
 	            {
@@ -397,12 +397,12 @@ namespace OpenMS
 	                {
 	                  // Compute the shift corresponding to a pair of elements.
 	                  Shift shift = shift_( (*model_map_)[*model_iter], (*scene_map_)[*scene_iter] );
-	
+
 	                  DPosition<2> tpwm = shift.position;
 	                  tpwm -= tbbe_min;
-	
+
 	                  DoubleReal  const & tq = shift.quality;
-	
+
 	                  // Compute the bucket index (the lowest of the four) for
 	                  // this shift.  Also compute the fractional part of
 	                  // the position within the bucket.
@@ -416,39 +416,39 @@ namespace OpenMS
 	                  }
 	                  DPosition<2> bucket_fraction_complement(1,1);
 	                  bucket_fraction_complement -= bucket_fraction;
-	
+
 	                  // Distribute the quality of the shift among the four neighboring buckets.
 	                  DoubleReal factor;
-	
+
 	                  factor = bucket_fraction_complement[Peak2D::RT] * bucket_fraction_complement[Peak2D::MZ];
 	                  tb( bucket_index[Peak2D::RT], bucket_index[Peak2D::MZ] ) += tq * factor;
-	
+
 	                  factor = bucket_fraction_complement[Peak2D::RT] * bucket_fraction[Peak2D::MZ];
 	                  tb( bucket_index[Peak2D::RT], bucket_index[Peak2D::MZ] + 1 ) += tq * factor;
-	
+
 	                  factor = bucket_fraction[Peak2D::RT] * bucket_fraction_complement[Peak2D::MZ];
 	                  tb( bucket_index[Peak2D::RT] + 1, bucket_index[Peak2D::MZ] ) += tq * factor;
-	
+
 	                  factor = bucket_fraction[Peak2D::RT] * bucket_fraction[Peak2D::MZ];
 	                  tb( bucket_index[Peak2D::RT] + 1, bucket_index[Peak2D::MZ] + 1 ) += tq * factor;
-	
+
 	                  ++number_of_considered_element_pairs_for_this_pair_of_buckets;
-	
+
 	                  if ( progress_dots &&
 	                       ! (number_of_considered_element_pairs_for_this_pair_of_buckets % progress_dots)
 	                     )
 	                  {
 	                    std::cout << 'H' << std::flush;
 	                  }
-	
+
 	                } // for scene_iter
 	              } // for model_iter
 	            } // for model_bucket_index_MZ
 	          } // for model_bucket_index_RT
 	        } // for scene_bucket_index_MZ
 	      } // for scene_bucket_index_RT
-	
-	
+
+
 	      // Optionally, write debug output as specified in param.
 	      if ( getParameters().exists("debug:dump_shift_buckets") )
 	      {
@@ -456,7 +456,7 @@ namespace OpenMS
 	        std::ofstream dump_file(dump_filename.c_str());
 	        dump_file << "# " << dump_filename << " generated " << QDateTime::currentDateTime().toString("yyyy-MM-dd hh:mm:ss").toStdString() << std::endl;
 	        dump_file << "# Shift buckets: xcoord ycoord quality xindex yindex" << std::endl;
-	
+
 	        for ( ShiftQualityMatrixType::ConstIterator iter = tb.begin(); iter != tb.end(); ++iter)
 	        {
 	          std::pair<UInt,UInt> row_col = tb.indexPair(iter-tb.begin());
@@ -473,30 +473,30 @@ namespace OpenMS
 	        dump_file << "# " << dump_filename << " EOF " << QDateTime::currentDateTime().toString("yyyy-MM-dd hh:mm:ss").toStdString() << std::endl;
 	      }
 	    }
-	
+
 	    /**
 	    	@brief Compute the shift.
-	
+
 	    	@note shift_buckets_ must have been calculated before.
 	    */
 	    void computeShift_(TransformationDescription& mapping)
 	    {
 	      Shift shift;
-	
+
 	      // Shorthands ...
 	      ShiftQualityMatrixType const & tb = shift_bucket_;
 	      DPosition<2> const & tbs = shift_bucket_size_;
 	      UInt const (&tbw)[2] = shift_bucket_window_;
-	
+
 	      // Find the transformation bucket with highest impact (quality).
 	      UInt tb_max_element_index = std::max_element(tb.begin(),tb.end()) - tb.begin();
 	      UInt tb_max_indices[2];
-	      tb_max_indices[Peak2D::RT] = tb.rowIndex(tb_max_element_index);
-	      tb_max_indices[Peak2D::MZ] = tb.colIndex(tb_max_element_index);
-	
+	      tb_max_indices[Peak2D::RT] = (int)tb.rowIndex(tb_max_element_index);
+	      tb_max_indices[Peak2D::MZ] = (int)tb.colIndex(tb_max_element_index);
+
 	      // Compute a weighted average of the shifts nearby the tb_max_element.
 	      //Shift result; // initially zero
-	
+
 	      DPosition<2> const& tbbe_min = shift_bounding_box_enlarged_.min();
 	      int tb_run_indices[2];
 	      for ( tb_run_indices[Peak2D::RT]  = std::max ( int (tb_max_indices[Peak2D::RT] - tbw[Peak2D::RT]), 0 );
@@ -525,20 +525,20 @@ namespace OpenMS
 	      {
 	        shift.position /= -shift.quality ;
 	      }
-				
+
 				mapping.setName("linear");
 	      mapping.setParam("slope",1.0);
 	      mapping.setParam("intercept",shift.position[0]);
 	    }
-	
+
 	    /**
 				@brief Compute the shift and similarity for a pair of elements; larger quality values are better.
-				
+
 				The returned value should express our confidence that one element might possibly be matched to the other.
-				
+
 				Currently this will just calculate the ratio of intensities, either
 				"left/right" or "right/left", such that a value between 0 and 1 is returned.
-				
+
 				@improvement Take the quality of the elements themselves into account, i.e., how good they fit to their model. (Eva)
 	    */
 	    Shift shift_( PointType const & left, PointType const & right ) const
@@ -556,42 +556,42 @@ namespace OpenMS
 	      }
 	      return shift;
 	    }
-			
+
 			///Pointer to the model map
 			const ConsensusMap* model_map_;
 			///Pointer to the scene map
 			const ConsensusMap* scene_map_;
-			
+
 	    /// Holds the bounding box of all input elements.
 	    DBoundingBox<2>  element_map_position_bounding_box_[2];
-	
+
 	    /// Holds the enlarged bounding box for all input elements.  It is larger  by about half of a bucket in all directions.
 	    DBoundingBox<2>  element_map_position_bounding_box_enlarged_[2];
-	
+
 	    /// Holds a bounding box for the input element intensities.
 	    DBoundingBox<1> element_map_intensity_bounding_box_[2];
-	
+
 	    /// Element indices are stored in theses buckets.
 	    ElementBucketMatrixType element_bucket_[2];
-	
+
 	    /// Diagonal size of each bucket.
 	    DPosition<2> element_bucket_size_;
-	
+
 	    /// Shifts are stored (summed up) in these buckets.
 	    ShiftQualityMatrixType shift_bucket_;
-	
+
 	    /// Holds a bounding box for all possible shift vectors.
 	    DBoundingBox<2> shift_bounding_box_;
-	
+
 	    /// Holds an enlarged bounding box for all shift vectors.  It is arger by about half of a bucket in all directions.
 	    DBoundingBox<2> shift_bounding_box_enlarged_;
-	
+
 	    /// Diagonal size of each bucket in shift_bucket_.
 	    DPosition<2> shift_bucket_size_;
-	
+
 	    /// Number of surrounding buckets of element indices to be considered when computing shifts.
 	    UInt element_bucket_window_[2];
-	
+
 	    /// Number of surrounding buckets of shift indices to be considered when computing shifts.
 	    UInt shift_bucket_window_[2];
   };
