+++ source/TRANSFORMATIONS/RAW2PEAK/OptimizePick.C
@@ -39,31 +39,31 @@ namespace OpenMS
 
 
 
-    // Print the computed signal 
+    // Print the computed signal
 			void printSignal(const gsl_vector* x,void* param, float resolution)
     {
 
-			std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (param) ->positions; 
-  		std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (param) ->peaks; 
+			std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (param) ->positions;
+  		std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (param) ->peaks;
       std::cout << "Printing Signal" << std::endl;
       if (resolution == 1.)
       {
-        // iterate over all points of the signal 
+        // iterate over all points of the signal
         for (size_t current_point = 0; current_point < positions.size(); current_point++)
         {
           double computed_signal     = 0.;
           double current_position    = positions[current_point];
 
-          // iterate over all peaks 
+          // iterate over all peaks
           for (size_t current_peak = 0; current_peak < peaks.size(); current_peak++)
           {
-            // Store the current parameters for this peak 
+            // Store the current parameters for this peak
             double p_height 		 = gsl_vector_get(x, 4*current_peak);
             double p_position    = gsl_vector_get(x, 4*current_peak + 3);
             double p_width  		 = (current_position <= p_position) ? gsl_vector_get(x, 4*current_peak + 1)
                                  : gsl_vector_get(x, 4*current_peak + 2);
 
-            // is it a Lorentz or a Sech - Peak? 
+            // is it a Lorentz or a Sech - Peak?
             if (peaks[current_peak].type == PeakShape::LORENTZ_PEAK)
             {
               computed_signal += p_height / (1. + pow(p_width * (current_position - p_position), 2));
@@ -78,23 +78,23 @@ namespace OpenMS
       }
       else
       {
-        // Compute step width 
+        // Compute step width
         float sw = (positions[1] - positions[0])/resolution;
         for (int i=0; i<positions.size() * resolution; i++)
         {
           double computed_signal     = 0.;
           double current_position    = positions[0] + i*sw;
 
-          // iterate over all peaks 
+          // iterate over all peaks
           for (size_t current_peak = 0; current_peak < peaks.size(); current_peak++)
           {
-            // Store the current parameters for this peak 
+            // Store the current parameters for this peak
             double p_height 		 = gsl_vector_get(x, 4*current_peak);
             double p_position    = gsl_vector_get(x, 4*current_peak + 3);
             double p_width  		 = (current_position <= p_position) ? gsl_vector_get(x, 4*current_peak + 1)
                                  : gsl_vector_get(x, 4*current_peak + 2);
 
-            // is it a Lorentz or a Sech - Peak? 
+            // is it a Lorentz or a Sech - Peak?
             if (peaks[current_peak].type == PeakShape::LORENTZ_PEAK)
             {
               computed_signal += p_height / (1. + pow(p_width * (current_position - p_position), 2));
@@ -113,7 +113,7 @@ namespace OpenMS
     }
 
 
-    // Evaluation of the target function for nonlinear optimization. 
+    // Evaluation of the target function for nonlinear optimization.
     int residual(const gsl_vector* x, void* params , gsl_vector* f)
     {
       // According to the gsl conventions, x contains the parameters to be optimized.
@@ -127,27 +127,27 @@ namespace OpenMS
       // instead.
       // The vector f is supposed to contain the result when we return from this function.
       // Note: GSL wants the values for each data point i as one component of the results vector
-				std::vector<DoubleReal>& signal = static_cast<OptimizePick::Data*> (params) ->signal; 
-				std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (params) ->positions; 
-				std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (params) ->peaks; 
+				std::vector<DoubleReal>& signal = static_cast<OptimizePick::Data*> (params) ->signal;
+				std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (params) ->positions;
+				std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (params) ->peaks;
         OptimizationFunctions::PenaltyFactors& penalties=static_cast<OptimizePick::Data*> (params) ->penalties;
-      // iterate over all points of the signal 
+      // iterate over all points of the signal
       for (size_t current_point = 0; current_point < positions.size(); current_point++)
       {
         double computed_signal     = 0.;
         double current_position    = positions[current_point];
         double experimental_signal = signal[current_point];
 
-        // iterate over all peaks 
+        // iterate over all peaks
         for (size_t current_peak = 0; current_peak < peaks.size(); current_peak++)
         {
-          // Store the current parameters for this peak 
+          // Store the current parameters for this peak
           double p_height 		 = gsl_vector_get(x, 4*current_peak);
           double p_position    = gsl_vector_get(x, 4*current_peak + 3);
           double p_width  		 = (current_position <= p_position) ? gsl_vector_get(x, 4*current_peak + 1)
                                : gsl_vector_get(x, 4*current_peak + 2);
 
-          // is it a Lorentz or a Sech - Peak? 
+          // is it a Lorentz or a Sech - Peak?
           if (peaks[current_peak].type == PeakShape::LORENTZ_PEAK)
           {
             computed_signal += p_height / (1. + pow(p_width * (current_position - p_position), 2));
@@ -166,7 +166,7 @@ namespace OpenMS
       double penalty_lwidth = penalties.lWidth;
       double penalty_rwidth = penalties.rWidth;
 
-      // iterate over all peaks again to compute the penalties 
+      // iterate over all peaks again to compute the penalties
       for (size_t current_peak = 0; current_peak < peaks.size(); current_peak++)
       {
         double old_position = peaks[current_peak].mz_position;
@@ -198,25 +198,25 @@ namespace OpenMS
       // Note: GSL expects the Jacobian as follows:
       // 	- each row corresponds to one data point
       // 	- each column corresponds to one parameter
-			//	std::vector<DoubleReal>& signal = static_cast<OptimizePick::Data*> (params) ->signal; 
-				std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (params) ->positions; 
-				std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (params) ->peaks; 
+			//	std::vector<DoubleReal>& signal = static_cast<OptimizePick::Data*> (params) ->signal;
+				std::vector<DoubleReal>& positions = static_cast<OptimizePick::Data*> (params) ->positions;
+				std::vector<PeakShape>& peaks = static_cast<OptimizePick::Data*> (params) ->peaks;
         OptimizationFunctions::PenaltyFactors& penalties=static_cast<OptimizePick::Data*> (params) ->penalties;
-      // iterate over all points of the signal 
+      // iterate over all points of the signal
       for (size_t current_point = 0; current_point < positions.size(); current_point++)
       {
         double current_position    = positions[current_point];
 
-        // iterate over all peaks 
+        // iterate over all peaks
         for (size_t current_peak = 0; current_peak < peaks.size(); current_peak++)
         {
-          // Store the current parameters for this peak 
+          // Store the current parameters for this peak
           double p_height 		 = gsl_vector_get(x, 4*current_peak);
           double p_position    = gsl_vector_get(x, 4*current_peak + 3);
           double p_width  		 = (current_position <= p_position) ? gsl_vector_get(x, 4*current_peak + 1)
                                : gsl_vector_get(x, 4*current_peak + 2);
 
-          // is it a Lorentz or a Sech - Peak? 
+          // is it a Lorentz or a Sech - Peak?
           if (peaks[current_peak].type == PeakShape::LORENTZ_PEAK)
           {
             double diff      = current_position - p_position;
@@ -332,7 +332,7 @@ void OptimizePick::optimize(std::vector<PeakShape>& peaks,Data& data)
     if (peaks.size() == 0)
       return;
 
-    int global_peak_number=0;
+    size_t global_peak_number=0;
     data.peaks.assign(peaks.begin(),peaks.end());
 
     gsl_vector *start_value=gsl_vector_alloc(4*data.peaks.size());
@@ -380,10 +380,10 @@ void OptimizePick::optimize(std::vector<PeakShape>& peaks,Data& data)
 
     gsl_multifit_fdfsolver_set(fit, &fit_function, start_value);
 
-    // initial norm 
+    // initial norm
     // std::cout << "Before optimization: ||f|| = " << gsl_blas_dnrm2(fit->f) << std::endl;
 
-    // Iteration 
+    // Iteration
     unsigned int iteration = 0;
     int status;
 
@@ -399,7 +399,7 @@ void OptimizePick::optimize(std::vector<PeakShape>& peaks,Data& data)
 #endif
       if (isnan(gsl_blas_dnrm2(fit->dx)))
         break;
- 
+
       // We use the gsl function gsl_multifit_test_delta to decide if we can finish the iteration.
       // We only finish if all new parameters deviates only by a small amount from the parameters of the last iteration
       status = gsl_multifit_test_delta(fit->dx, fit->x, eps_abs_, eps_rel_);
@@ -418,17 +418,17 @@ void OptimizePick::optimize(std::vector<PeakShape>& peaks,Data& data)
 
     //		OptimizationFunctions::printSignal(fit->x, 5.,param);
 
-    // iterate over all peaks and store the optimized values in peaks 
+    // iterate over all peaks and store the optimized values in peaks
    for (size_t current_peak = 0; current_peak <data.peaks.size(); current_peak++)
     {
-      // Store the current parameters for this peak 
+      // Store the current parameters for this peak
       peaks[global_peak_number+current_peak].height 		 = gsl_vector_get(fit->x, 4*current_peak);
       peaks[global_peak_number+current_peak].mz_position = gsl_vector_get(fit->x, 4*current_peak + 3);
       peaks[global_peak_number+current_peak].left_width  = gsl_vector_get(fit->x, 4*current_peak + 1);
       peaks[global_peak_number+current_peak].right_width = gsl_vector_get(fit->x, 4*current_peak + 2);
 
       //	compute the area
-      // is it a Lorentz or a Sech - Peak? 
+      // is it a Lorentz or a Sech - Peak?
       if (peaks[global_peak_number+current_peak].type == PeakShape::LORENTZ_PEAK)
       {
         PeakShape p = peaks[global_peak_number+current_peak];
